Move subcomponents to separate files, add default export
