+++ packages/components/src/composite/index.tsx
@@ -22,142 +22,14 @@ import { useMemo, forwardRef } from '@wordpress/element';
  * Internal dependencies
  */
 import type { WordPressComponentProps } from '../context';
-import { CompositeContext, useCompositeContext } from './context';
-import type {
-	CompositeStoreProps,
-	CompositeProps,
-	CompositeGroupProps,
-	CompositeGroupLabelProps,
-	CompositeItemProps,
-	CompositeRowProps,
-	CompositeHoverProps,
-	CompositeTypeaheadProps,
-} from './types';
-
-/**
- * Creates a composite store.
- *
- * @example
- * ```jsx
- * import { Composite, useCompositeStore } from '@wordpress/components';
- *
- * const store = useCompositeStore();
- * <Composite store={store}>
- *   <Composite.Item>Item</Composite.Item>
- *   <Composite.Item>Item</Composite.Item>
- *   <Composite.Item>Item</Composite.Item>
- * </Composite>
- * ```
- */
-export function useCompositeStore( {
-	focusLoop = false,
-	focusWrap = false,
-	focusShift = false,
-	virtualFocus = false,
-	orientation = 'both',
-	rtl = false,
-	...props
-}: CompositeStoreProps = {} ) {
-	return Ariakit.useCompositeStore( {
-		focusLoop,
-		focusWrap,
-		focusShift,
-		virtualFocus,
-		orientation,
-		rtl,
-		...props,
-	} );
-}
-
-const Group = forwardRef<
-	HTMLDivElement,
-	WordPressComponentProps< CompositeGroupProps, 'div', false >
->( function CompositeGroup( props, ref ) {
-	const context = useCompositeContext();
-	return (
-		<Ariakit.CompositeGroup
-			store={ context?.store }
-			{ ...props }
-			ref={ ref }
-		/>
-	);
-} );
-Group.displayName = 'Composite.Group';
-
-const GroupLabel = forwardRef<
-	HTMLDivElement,
-	WordPressComponentProps< CompositeGroupLabelProps, 'div', false >
->( function CompositeGroupLabel( props, ref ) {
-	const context = useCompositeContext();
-	return (
-		<Ariakit.CompositeGroupLabel
-			store={ context?.store }
-			{ ...props }
-			ref={ ref }
-		/>
-	);
-} );
-GroupLabel.displayName = 'Composite.GroupLabel';
-
-const Item = forwardRef<
-	HTMLButtonElement,
-	WordPressComponentProps< CompositeItemProps, 'button', false >
->( function CompositeItem( props, ref ) {
-	const context = useCompositeContext();
-	return (
-		<Ariakit.CompositeItem
-			store={ context?.store }
-			{ ...props }
-			ref={ ref }
-		/>
-	);
-} );
-Item.displayName = 'Composite.Item';
-
-const Row = forwardRef<
-	HTMLDivElement,
-	WordPressComponentProps< CompositeRowProps, 'div', false >
->( function CompositeRow( props, ref ) {
-	const context = useCompositeContext();
-	return (
-		<Ariakit.CompositeRow
-			store={ context?.store }
-			{ ...props }
-			ref={ ref }
-		/>
-	);
-} );
-Row.displayName = 'Composite.Row';
-
-const Hover = forwardRef<
-	HTMLDivElement,
-	WordPressComponentProps< CompositeHoverProps, 'div', false >
->( function CompositeHover( props, ref ) {
-	const context = useCompositeContext();
-	return (
-		<Ariakit.CompositeHover
-			store={ context?.store }
-			{ ...props }
-			ref={ ref }
-		/>
-	);
-} );
-Hover.displayName = 'Composite.Hover';
-
-const Typeahead = forwardRef<
-	HTMLDivElement,
-	WordPressComponentProps< CompositeTypeaheadProps, 'div', false >
->( function CompositeTypeahead( props, ref ) {
-	const context = useCompositeContext();
-	return (
-		<Ariakit.CompositeTypeahead
-			store={ context?.store }
-			{ ...props }
-			ref={ ref }
-		/>
-	);
-} );
-Typeahead.displayName = 'Composite.Typeahead';
+import { CompositeContext } from './subcomponents/context';
+import { CompositeGroup } from './subcomponents/group';
+import { CompositeGroupLabel } from './subcomponents/group-label';
+import { CompositeHover } from './subcomponents/hover';
+import { CompositeItem } from './subcomponents/item';
+import { CompositeRow } from './subcomponents/row';
+import { CompositeTypeahead } from './subcomponents/typeahead';
+import type { CompositeProps } from './types';
 
 /**
  * Renders a widget based on the WAI-ARIA [`composite`](https://w3c.github.io/aria/#composite)
@@ -222,7 +94,7 @@ export const Composite = Object.assign(
 		 * </Composite>
 		 * ```
 		 */
-		Group,
+		Group: CompositeGroup,
 		/**
 		 * Renders a label in a composite group. This component must be wrapped with
 		 * `Composite.Group` so the `aria-labelledby` prop is properly set on the
@@ -242,7 +114,7 @@ export const Composite = Object.assign(
 		 * </Composite>
 		 * ```
 		 */
-		GroupLabel,
+		GroupLabel: CompositeGroupLabel,
 		/**
 		 * Renders a composite item.
 		 *
@@ -258,7 +130,7 @@ export const Composite = Object.assign(
 		 * </Composite>
 		 * ```
 		 */
-		Item,
+		Item: CompositeItem,
 		/**
 		 * Renders a composite row. Wrapping `Composite.Item` elements within
 		 * `Composite.Row` will create a two-dimensional composite widget, such as a
@@ -283,7 +155,7 @@ export const Composite = Object.assign(
 		 * </Composite>
 		 * ```
 		 */
-		Row,
+		Row: CompositeRow,
 		/**
 		 * Renders an element in a composite widget that receives focus on mouse move
 		 * and loses focus to the composite base element on mouse leave. This should
@@ -304,7 +176,7 @@ export const Composite = Object.assign(
 		 * </Composite>
 		 * ```
 		 */
-		Hover,
+		Hover: CompositeHover,
 		/**
 		 * Renders a component that adds typeahead functionality to composite
 		 * components. Hitting printable character keys will move focus to the next
@@ -321,7 +193,7 @@ export const Composite = Object.assign(
 		 * </Composite>
 		 * ```
 		 */
-		Typeahead,
+		Typeahead: CompositeTypeahead,
 		/**
 		 * The React context used by the composite components. It can be used by
 		 * to access the composite store, and to forward the context when composite
@@ -339,3 +211,8 @@ export const Composite = Object.assign(
 		Context: CompositeContext,
 	}
 );
+
+export { useCompositeStore } from './subcomponents/store';
+
+// Works best with Storybook's prop extraction
+export default Composite;
