+++ packages/components/src/composite/subcomponents/context.ts
@@ -6,7 +6,7 @@ import { createContext, useContext } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import type { CompositeContextProps } from './types';
+import type { CompositeContextProps } from '../types';
 
 export const CompositeContext =
 	createContext< CompositeContextProps >( undefined );
